Подключила дополнительные модули. Перенесла туда 2 функции
