+++ js/main.js
@@ -1,11 +1,3 @@
-
-function getRandomInteger(min, max) {
-  return Math.floor(Math.random() * (max - min + 1) ) + min;
-}
-getRandomInteger ();
-function getLengthString(stringNumber, maxLength ) {
-  let maxStringLength = 140;
-
-  return (maxLength < maxStringLength);
-}
-getLengthString();
+import './util.js';
+import './data.js';
+import './basic.js';
