+++ js/util.js
@@ -1,0 +1,13 @@
+function getRandomInteger(min, max) {
+  return Math.floor(Math.random() * (max - min + 1) ) + min;
+}
+getRandomInteger();
+function getLengthString(stringNumber, maxLength ) {
+  let maxStringLength = 140;
+
+  return (maxLength < maxStringLength);
+}
+getLengthString();
+
+export{getRandomInteger};
+export{getLengthString};
